+++ androidx/src/main/java/me/panpf/androidx/app/Activityx.java
@@ -35,6 +35,7 @@ import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.List;
 
+import me.panpf.androidx.content.Contextx;
 import me.panpf.javax.lang.Classx;
 
 @SuppressWarnings({"WeakerAccess"})
@@ -153,12 +154,6 @@ public class Activityx {
     }
 
 
-    @NonNull
-    public static Context appContext(@NonNull Activity activity) {
-        return activity.getApplicationContext();
-    }
-
-
     /* ************************************* canStart ***************************************** */
 
 
@@ -228,7 +223,7 @@ public class Activityx {
      */
     public static void start(@NonNull android.support.v4.app.Fragment fragment,
                              @NonNull Class<? extends Activity> clazz, @Nullable Bundle args) {
-        Intent intent = new Intent(Fragmentx.requireContext(fragment), clazz);
+        Intent intent = new Intent(Contextx.requireContext(fragment), clazz);
         if (args != null) intent.putExtras(args);
         fragment.startActivity(intent);
     }
@@ -246,7 +241,7 @@ public class Activityx {
      */
     public static void start(@NonNull android.app.Fragment fragment,
                              @NonNull Class<? extends Activity> clazz, @Nullable Bundle args) {
-        Intent intent = new Intent(Fragmentx.requireContext(fragment), clazz);
+        Intent intent = new Intent(Contextx.requireContext(fragment), clazz);
         if (args != null) intent.putExtras(args);
         fragment.startActivity(intent);
     }
@@ -351,7 +346,7 @@ public class Activityx {
      */
     public static boolean safeStart(@NonNull android.support.v4.app.Fragment fragment,
                                     @NonNull Class<? extends Activity> clazz, @Nullable Bundle args) {
-        Intent intent = new Intent(Fragmentx.requireContext(fragment), clazz);
+        Intent intent = new Intent(Contextx.requireContext(fragment), clazz);
         if (args != null) intent.putExtras(args);
         return safeStart(fragment, intent);
     }
@@ -369,7 +364,7 @@ public class Activityx {
      */
     public static boolean safeStart(@NonNull android.app.Fragment fragment,
                                     @NonNull Class<? extends Activity> clazz, @Nullable Bundle args) {
-        Intent intent = new Intent(Fragmentx.requireContext(fragment), clazz);
+        Intent intent = new Intent(Contextx.requireContext(fragment), clazz);
         if (args != null) intent.putExtras(args);
         return safeStart(fragment, intent);
     }